requirements.txt
fastapi
uvicorn
python-dotenv
telethon
google-generativeai
# Added for classifier service (no new external libs beyond FastAPI/uvicorn)